remove junk from the S2B previously done, and started the implementation of B2ISBipolar
