--- src/main/scala/project/B2IS.scala
@@ -1,6 +1,0 @@
-package project
-
-import chisel3._
-
-// Binary to Integer Stochastic
-class B2IS extends Module {}
