+++ src/main/scala/project/SCNN.scala
@@ -1,12 +1,12 @@
 package project
-import project.B2S
+import project.B2SUnipolar
 
 import chisel3._
 import _root_.circt.stage.ChiselStage
 
 object StochasticNeuralNetwork extends App {
   ChiselStage.emitSystemVerilogFile(
-    new B2S,
+    new B2SUnipolar,
     args = Array(
       "--target-dir",
       "generated/project/"
